it works by iterating over the meals list reptdly!
prints out a bunch of garbage, AND prioritizes the first meals given in the list for who will get any remainder friends, as in, if there are 9 people (as there are this weekend :) ) then the list of those preparing brunch will automatically have 3 and the rest only 2.  I wonder if there is a way to randomize (there's a hint :) ) who would get the remainder person.  Probably something like if k is not % len(j): meals.append(random.j) or something like that??  oh jeez that looks complicated.

and it errors out at the end because I haven't found a good way for it to deal with running out of people.

BUT STILL!  IT WORKS & PRINTS THE LEGIT MEAL FIXERS OUT, and I cut down 20 lines!